Solucion de errores, falta conectar el posmant genera errror al conectar

+++ src/main/java/com/bogallery/bogalleryApp/service/InscripcionService.java
@@ -5,11 +5,13 @@ import com.bogallery.bogalleryApp.entities.Inscripcion;
 import java.util.List;
 import java.util.Optional;
 
+
+
 public interface InscripcionService {
 
     public List<Inscripcion> findAll() throws Exception;
 
-    public Inscripcion findById(int id);
+    public Inscripcion findById(Long id);
 
     public void create(Inscripcion inscripcion);
     public void update(Inscripcion inscripcion);
