+++ src/main/java/com/bogallery/bogalleryApp/service/PagoService.java
@@ -9,7 +9,7 @@ public interface PagoService {
 
     public List<Pago> findAll() throws Exception;
 
-    public Pago findById(int id);
+    public Pago findById(Long id);
 
     public void create(Pago pago);
     public void update(Pago pago);
